backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
distlib==0.3.8
Django==5.0.4
django-colorfield==0.11.0
django-credit-cards==0.4.1
django-phonenumber-field==7.3.0
django-tenants==3.6.1
djangorestframework==3.15.1
filelock==3.13.4
Markdown==3.6
phonenumberslite==8.13.35
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
virtualenv==20.25.2
